requirements.txt
azure-functions
beautifulsoup4
requests
